Update Dockerfile.rtlsdr (#40)
* Add in pkg-config to stop rtlsdr config whinging about it
* Switch from tagged branch to master

My theory on what is happening is that the tag we are currently using is somehow borked and enabling zero-copy by default, or otherwise doing something that causes programs using lib-rtlsdr to be high in CPU usage. Moving to `master` appears to fix this.

I don't know what potential negative impact this will have, but my assumption is that code that ends up in master on their end is vetted and okay to use.